Improved test code organization
