--- test/__main__.py
@@ -1,19 +1,0 @@
-#!/usr/bin/env python
-
-import sys
-import os
-
-try:
-	import itcsimlib
-except ImportError:
-	sys.path.append(os.path.abspath("."))
-
-from base import *
-from model import *
-from utilities import *
-from drakon import *
-from massspec import *
-from trap import *
-
-if __name__ == '__main__':
-	unittest.main()
